+++ frontend/app/components/prints/FileList.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 /**
  * A reusable file list component, displays files in a simple list, includes:
- *  - File name, category and date displayed per file //TODO atm only year is displayed bc no actual data yet
+ *  - File name, category and date displayed per file
  *
  * Usage:
  * <PrintsFileList
@@ -10,13 +10,13 @@
  * />
  *
  * Example files:
- * const files: PrintsFile[] = [
- *    { id: 1, name: 'AFFICHE-FESTIVAL-2025.PDF', year: 2025, image: null },
- *    { id: 2, name: 'AFFICHE-VIDEODROOM-2024.PDF', year: 2024, image: null },
+ * const files: PrintItemView[] = [
+ *    { id: 1, titel: 'AFFICHE-FESTIVAL-2025.PDF', description: '', url: '', created_at: '...', updated_at: '...' },
+ *    { id: 2, titel: 'AFFICHE-VIDEODROOM-2024.PDF', description: '', url: '', created_at: '...', updated_at: '...' },
  * ]
  */
 
-import { FileText, ChevronLeft, ChevronRight } from "lucide-vue-next";
+import { ChevronLeft, ChevronRight } from "lucide-vue-next";
 import type { PrintItemView } from "@repo/common";
 
 interface Props {
@@ -24,10 +24,9 @@ interface Props {
   files: PrintItemView[];
 }
 const props = defineProps<Props>();
-const { t, locale } = useI18n();
+const { t } = useI18n();
 
 // file logic
-const openFile = (src: string) => window.open(src, "_blank");
 const emit = defineEmits<{
   (e: "delete", file: PrintItemView): void;
 }>();
@@ -50,8 +49,6 @@ function goToPage(page: number) {
 }
 
 // constants
-const rowBase =
-  "group relative flex items-center gap-4 px-4 py-3 border-t border-border bg-card hover:bg-card-hover transition-colors duration-150 cursor-pointer";
 const list = computed(() =>
   totalPages.value > 1 ? { minHeight: `${ITEMS_PER_PAGE * 63}px` } : {},
 ); // 62 is height of one row (approximately)
@@ -67,59 +64,13 @@ const chevronButton =
       class="rounded-lg border border-border overflow-hidden"
       :style="list"
     >
-      <div
+      <PrintsFileListItem
         v-for="file in visibleFiles"
         :key="file.id"
-        :class="rowBase"
-        @click="file.url ? openFile(file.url) : undefined"
-      >
-        <!-- Icon -->
-        <div
-          class="w-9 h-9 rounded-md bg-muted flex items-center justify-center shrink-0"
-        >
-          <FileText :size="18" class="text-muted-foreground" />
-        </div>
-
-        <!-- File info -->
-        <div class="flex-1 min-w-0">
-          <p
-            class="text-[13px] font-bold truncate group-hover:text-accent transition-colors duration-150"
-          >
-            {{ file.titel }}
-          </p>
-
-          <p
-            class="text-[10px] text-muted-foreground uppercase tracking-widest mt-0.5"
-          >
-            {{ category }} |
-            <span v-if="file.created_at">
-              {{
-                new Date(file.created_at).toLocaleDateString(locale, {
-                  day: "numeric",
-                  month: "long",
-                  year: "numeric",
-                })
-              }}
-            </span>
-          </p>
-        </div>
-
-        <!-- Buttons -->
-        <div class="flex items-center gap-2 shrink-0" @click.stop>
-          <AdminDownloadButton
-            v-if="file.url"
-            :label="t('prints.download')"
-            :size="37"
-            :src="file.url"
-            :name="file.url"
-          />
-          <AdminDeleteButton
-            :label="t('prints.delete')"
-            :size="37"
-            @click="emit('delete', file)"
-          />
-        </div>
-      </div>
+        :file="file"
+        :category="category"
+        @delete="emit('delete', file)"
+      />
     </div>
 
     <!-- Pagination -->
